Add Icons to Heading Item When Not in Source Mode
